requirements.txt
boto3==1.21.16
aws_lambda_powertools==1.25.3
Pillow==9.0.1
tweepy==4.7.0
requests==2.27.1
# dotenv